norm done, check size 100

+++ linked_list.c
@@ -6,55 +6,55 @@
 /*   By: mishin <mishin@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/02 18:45:54 by mishin            #+#    #+#             */
-/*   Updated: 2021/07/14 17:07:16 by mishin           ###   ########.fr       */
+/*   Updated: 2021/07/25 12:45:54 by mishin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-t_ll    *ll_new(int	data)
+t_ll	*ll_new(long long data)
 {
-        t_ll    *ret;
+	t_ll	*ret;
 
-        ret = (t_ll *)malloc(sizeof(t_ll));
-        if (!ret)
-                return (NULL);
-        ret->data = data;
-        ret->next = NULL;
-        return (ret);
+	ret = (t_ll *)malloc(sizeof(t_ll));
+	if (!ret)
+		return (NULL);
+	ret->data = data;
+	ret->next = NULL;
+	return (ret);
 }
 
-void    ll_push(t_ll **head, t_ll *new)
+void	ll_push(t_ll **head, t_ll *new)
 {
-        t_ll    *cur;
+	t_ll	*cur;
 
-        if (!head)
-                return ;
-        if (!*head)
-        {
-                *head = new;
-                return ;
-        }
-        cur = *head;
-        while (cur->next)
-                cur = cur->next;
-        cur->next = new;
+	if (!head)
+		return ;
+	if (!*head)
+	{
+		*head = new;
+		return ;
+	}
+	cur = *head;
+	while (cur->next)
+		cur = cur->next;
+	cur->next = new;
 }
 
-int             ll_clear(t_ll **head)
+int	ll_clear(t_ll **head)
 {
-        t_ll    *cur;
-        t_ll    *tmp;
+	t_ll	*cur;
+	t_ll	*tmp;
 
-        if (!head)
-                return (-1);
-        cur = *head;
-        while (cur)
-        {
-                tmp = cur->next;
-                free(cur);
-                cur = tmp;
-        }
-        *head = NULL;
-        return (-1);
+	if (!head)
+		return (-1);
+	cur = *head;
+	while (cur)
+	{
+		tmp = cur->next;
+		free(cur);
+		cur = tmp;
+	}
+	*head = NULL;
+	return (-1);
 }
